Refs #85. Fixed some issues during test.
